bad_argument! macro
`bad_argument!()` will create a `BadArgument` that records the `file`,
`line`, and `column` of the caller, so that it is easier to debug why
a `BadArgument` was generated.  This is mostly for debugging test
failures quickly as it is only used for `DebugInProcess` used for
formatting in `assert_*_in_process!` macros.