:erlang.error/1
`:erlang.error(reason)` raises an error class exception with term
reason.  `:erlang.error(:badarg)` matches the behavior of "bad argument"
from BIFs in BEAM, so to support this equivalence, BadArgument has been
replaced with the more general Exception. It has a Class, which is one
of Error, Exit, or Throw to mirror the classes documented in
http://erlang.org/doc/reference_manual/errors.html.  The reason is a
Term.

BadArgument struct is replaced with Exception having Class Error and
reason, `:badarg`.  The `assert_bad_argument!` and `bad_argument!`
macros remain, but now there is also the more general `assert_error!`
and `error!` macros.  May methods had to be changed to accept the `&mut
Process` in case they need to create the error's reason Term.