Global atom table
Makes atom look up faster as it doesn't need to go through the process
to the environment to the atom table.

Removes a lot of the need to pass the process, so changes a lot of APIs.

Because the APIs changed, so much I added helpers for the tests for
redundant checks like "errors badarg", which shrunk those tests to
only the single value setup.