erlang::start_timer_3
start and timeout timers.  Timers are timed out by calling
`timer::timeout`.  There is hierarchy of timers per scheduler.  As
schedulers don't exist right now, the timer hierarchies are thread
local.

The `timer::Hierarchy` is based on the timer hiearchy used in BEAM and
described in
https://github.com/erlang/otp/blob/759ec896d7f254db2996cbb503c1ef883e6714b0/erts/emulator/beam/time.c#L21-L171.

* `at_once`
* `soon`
* `later`
* `long_term`

Although `erlang::start_timer_3` uses relative milliseconds, the timers
and `Hierarchy` use absolute monotonic time in milliseconds.  As such,
it is possible to start a timer that is before the current monotonic
timer in millisecond, which are put in the `at_once` min-heap to timeout
the next time `timer::timeout` is called.

Timers set to timeout in the next 2¹⁴ milliseconds (~16 seconds) are stored in
`soon` with one slot per millisecond.

Timers set to timeout between 2¹⁴ milliseconds and 2²⁸ milliseconds (~74
hours) are stored in `later` with one slot per 2¹³ (~8 seconds, half the
total milliseconds in `soon`).  Whenever `soon`'s `slot_index` advances, it is
checked if any timers in `later`'s `slot_index` should be transfered to
`soon`.  Timers are stored in a `min-heap` so as soon as the top of the
slot does not need to be transferred, no more timers need to checked
in the slot.  If the `soon`'s domain ever extends beyond the last millisecond
in the current `later` `slot_index`, `later`'s slot advances.

Timers that are beyond `later` are store in a single min-heap,
`long_term`.  Whenever `later`'s `slot_index` advances, it is checked if any
timers in `long_term` should be transferred to `later`.  Because of the
min-heap, as soon as the top of `later` does not need to be transferred,
no more timers need to be checked.

For testing, `soon` and `later` are restricted to 4 slots, the lowest
power of two to have a start (`0`), middle (`1`, `2`), and end ('3`)
index.