:erlang.cancel_timer/2
There is no asynchronous method of sending or receiving messages between
the timer hierarchies as placing an at once timer in the hierarchy would
require taking the same lock as removing the cancelled timer.  Instead,
only the effects of `async: true` are replicated: the `cancel_timer`
messages are sent to self when `info: true` or no `info` option is set.

Fixes a race condition in `:erlang.cancel_timer/1` tests where the
Scheduler could be dropped in `different_thread` before the timer was
canceled again, leading to the scheduler ID being reused and assigned to
a different timeout.  Fixed with an additional barrier wait.  This is
a test-only problem.  Outside of tests, schedulers won't shutdown once
started.