Convert www from submodule to directory
Using the Rust WASM book tools, the `www` directory is added as a
submodule, but that doesn't make sense because we want edits in our
repo, so delete the submodule setup and check in the local version of my
code.  I only noticed this when switching to `master` and back to `develop` as
the directory changed and I had to `rm -rf examples` to get the can't
switch branches error from `git`.