Initial implementation of term type scheme and garbage collection
The term type scheme is essential to the implementation of memory
management in the runtime, namely garbage collection, so this PR has
both. The scheme is broken down in three main parts:

- The type `Term` which is a very thin layer over tagged usize/ptr
values, this is what is used primarily within the garbage collector
- The type `TypedTerm` which is a struct enum type that is a richer
version of `Term` in that it has reified "real" values from the raw term
and then wrapped it in an enum that can be treated as something akin to
an "Any" type, i.e. it handles generically dispatching
equality/comparison operations, and is used to bundle up all the
different "real" types into a single generic one. This is what should be
used in most runtime functions when working with arbitrary terms.
- Finally, each variant of `TypedTerm` is really a wrapper for some
"real" type that handles the details of implementing that type, e.g.
integers, floats, binaries, etc. This is where the bulk of the logic for
implementing runtime support functions should live, leaving builtins to
simply handle converting a `Term` to a `TypedTerm`, verifying argument
types, and then dispatching to the runtime support functions via the wrapped values