Fix some issues with garbage collection/memory layout
It was missed, and turns out to be very important, that boxed terms,
with few exceptions, need to carry their arityval in their header, which
describes to the garbage collector how much data (in words) follows the
header term itself. When incorrect, the garbage collector will enter an
undefined state, as it potentially scans raw data as if they are valid
Terms. Most changes have to do with making sure these are correct.

Additionally, work was done to properly handle the virtual binary heap,
and off-heap fragments during garbage collection, both by factoring in
their sizes, as well as sweeping them for garbage when appropriate.

As of this commit, there are now two GC tests implemented, one that
performs a full sweep, and one that performs a minor collection. Two
more WIP tests are stubbed out here, but will follow shortly.