Return Boxed<Term> inside of TypedTerm::Boxed
Fix SEGV when using BigIntegers by having TypedTerm::Boxed contain
`Boxed<Term>` instead of `Term`.  It previously included the `Term` to
the tip of the boxed type, but then used that's Term's address, which
was the address of the copy when calling `to_typed_term` to get the
unboxed type.  Using `Boxed<Term` ensures the original address of the
tagged pointer is preserved when `to_typed_term` needs it for the
converting the header to the unboxed type.

This mimics the need to `unbox_reference<Term>` before `unbox_reference`
to the unboxed type in the previous system.