Term::is_runtime and Term::has_no_arity
is_runtime returns true if the type is allowed to be passed in from the
runtime, so immediates, boxed, and non-empty lists, but not headers
or move markers.

Term::has_no_arity returns true for the same types, but it is because
for code that does ptr math, the offset is just 1 instead of 1 + arity
because immediates are 1 term wide while boxed and non-empty list are
pointer types.