Fix HeapAlloc for HeapFragment
alloc incorrectly checked the available space between the base `data`
address and the current `top`, but when there is nothing allocated in
the heap yet these are equal, so `alloc` would always fail.

Calculate the `available` space as the distance between the current
`top` and the `max_top` (`data + size`).