Return `top` instead of `new_top` from HeapFragment alloc
Fix overwrites in ``:timeout` tuples due to `alloc` returning the
`new_top` instead of old `top`.  Rename `max_top` to `top_limit`
because `top` should never be returned if it is at `top_limit`.

As thought experiments, if the HeapFragment had `size` `0`, `top_limit`
and `top` would be the same and `AllocErr` should return.  If `size` is
`1`, then `top_limit = top + 1` and `1` word should be allocated by
returning `top` and then incrementing `self.top`, so the next call fails
with `AllocErr`.