Fix argument order in list_to_bitstring_1
Missed reordering the arguments when the parameters to
subbinary_from_original changed.  Renamed variables to match subbinary
fields to make the call clearer as with the old names the bit_offset
local was being used for the final partial_byte_bit_count instead of
the bit_offset, which was part of the cause of this being missed.