lumen_runtime::future::run_until_ready
Allow a process to be spawned that runs a function being placed in the child
process using `place_frame_With_arguments` and return the value to caller.
The current thread's Scheduler is run until the child process exits or
the Future is ready.

The returned Ready keeps both the `exception::Result` alive and the