Refactor garbage collector
The previous implementation of heaps/GC was difficult to test for a
variety of reasons, but most significantly due to how intertwined all
the pieces were - a fair number of "private" APIs had to be semi-public
to allow the GC to access the guts of the process heap, and
understanding which parts were involved in any one part was not very
clear.

Another major issue was the code duplication of extremely fragile memory
management code in the GC. A lot of duplicated effort around calculating
offsets to the next term, what to do when encountering specific kinds of
terms, and how to move those terms from one heap to another. In
addition, many of these places were not guaranteeing that things like
alignment were being upheld during moves.

This refactoring completely restructures the GC to provide the
following:

- The ability to test against non-process based heaps to help isolate
bugs when encountered
- Splits up core algorithmic details from process-specific setup/tear
down details, i.e. setting process flags, gathering roots, etc. This
allows the core garbage collector to be more easily tested, and then
consumed as a component without worrying that details are missed in
different GC implementations
- Uses the previously introduced HeapIter trait to walk heaps uniformly
- The ability to test individual components of the GC, such as moving
terms, and even moving specific kinds of terms (see the
`Sweeper`/`Sweepable` traits)
- Introduces a CollectionType trait which encompasses algorithmic differences in
different modes of a garbage collector, i.e. minor or full.
- Introduces a GarbageCollector trait to allow different GC
implementations to be used, primarily for testing
- Updates the RootSet to use Boxed to wrap terms, ensuring we don't copy
terms unintentionally
- Adds new tests, and updates existing tests