Implement compilation pipeline w/linking
- This change wires up the full compiler pipeline, from calculating the
inputs to compile, to codegen, generation of object files, and finally,
linking those objects into an executable or library.
- The compiler is parallelized:
- The compiler runs a thread pool of NUM_CPUS size, and each compilation
unit (i.e. each input file and its derivative artifacts) is compiled
entirely on one thread pool. This ensures that we don't get unintended
sharing of the LLVM/MLIR contexts, which I don't believe are
thread-safe.
- Likewise, the queries for the incremental engine that need anything
thread-local, are given the current thread id as part of the query, so
that the results are cached on a per-thread basis.
- Currently, I've marked a few of the key codegen types (the MLIR
context, LLVM context, MLIR/LLVM modules) as send/sync, which is safe as
long as we never access them from the wrong thread, which we currently
guarantee. Ideally though, we'll want something less prone to mistakes.
- Code generation is not yet complete (i.e. lowering from EIR to MLIR),
that is coming in a following commit.
- We ship LLD as part of the compiler, but it is not used on macOS by
default, as the backend for macOS is reportedly unmaintained and only
works on simple programs. However, it can be used by passing `-C
linker=lumen-lld`, and will be used by default on supporting platforms;
all other platforms use the system linker.