Initial sketch of lowering from EIR to MLIR
This commit mostly contains the skeleton of the lowering:

- The types in Rust which provide a safe API for consumption from the
compiler, and which manage references to the builder implemented in C++
- The outline of how module generation will proceed:
  - For each module,
  - For each function in that module,
    - Gather all of the constant atoms for atom table generation
    - Perform lowering analysis via libeir_lowerutils
    - Generate function(s) (i.e. if there are lifted functions/closures)
      - Declare function prototype
      - Define function body
        - For the function entry, generate the entry block which
          handles stack frame management and other setup
        - For each block in the function scope, generate that block
  - For each block in each function body,
    - Get the block arguments
    - Declare the block
    - Determine the block operation kind
    - Generate the implementation for that operation

The builder in Rust acts as the driver for the C++ builder, it
controls how the lowering proceeds, and translates data to C++
for that builder to consume.

Now that this compiles, what remains is just implementing the details of
each operations lowering, which for most operations is very
straightforward.