[compiler] start removing unnecessary implicit block arg handling
The origin translation to MLIR tried to force all block arguments to be
explicit, even when blocks only have one predecessor. This resulted in
some bugs encountered during lowering where the number and type of block
arguments did not match up with what was expected. This commit removes
the calculation of implicit block arguments, and instead uses some of
the same information to determine when a target block has arguments
expected to be filled by the operation in the predecessor block.

This commit may be too overzelous in removing block arguments, but I'd
rather go too far in removing it, and progressively add back what is
needed, rather than keep around a lot of extra useless code that makes
maintenance harder. There is enough debugging in place to make it
immediately obvious when things go wrong.