[codegen] bug fixes, cleanup, general enhancements
This commit blends together a bunch of changes since bug fixes led to
discovering more bugs, which led to more fixes/changes, etc.

A major change here was breaking up the original contents of
ConvertEIRToLLVM.cpp into subcategories of conversions. There is no hard
and fast rules on what the categories need to be, but I mostly
categorized them based on their fit as it is today. We can always
reorganize in the future.

There are some changes for future work in here, around closures, but
that can be mostly disregarded for the moment. The major
fixes/improvements are around the handling of constants and creation of
cons cells/tuples, as well as handling of casts. There is a large chunk
of commented out code in the cast lowering right now as I work support
for casts back in slowly after removing a bunch of handling - the idea
being that if we don't need it, we can get rid of it. Unfortunately
without the test suite up and running, its too early to do that just
yet.