Resource type name
Debugging Resource errors always required additional debug statements because the Any type only contains the type id, which can't be translated back to the type name since the correspondence isn't recorded.

To record the name of the type, `std::any::type_name` is used, but to get the original type, the `resource` functions can't take `Box<dyn Any>` because the type is already lost by then and all `type_name` would be `Box<dyn Any>`, so change the type of `value` to be generic.

This does mean that all previous callers that passed in a `Box::new` need to be changed so that they have a helpful type.