Fix spawn/1 not triggering badarity
Fixes #543

Spawn apply/2 when spawning a closure, so that apply/2 can raise badarity in the child process.  Because the code in trap_exceptions and trap_exceptions_impl only allows 1 argument to be passed and it, `erlang:apply/2` can be called directly.  Instead, a new `lumen:apply_apply_2` is created, which is passed a single argument, which is `[function, arguments]`.  That list is type checked and then `apply(function, arguments)` is called.