Timeouts don't call __lumen_builtin_receive_done
If `__lumen_builtin_receive_wait` returns `ReceiveState::Timeout`, `__lumen_builtin_receive_done` is not called, so if `ReceiveContext::should_time_out()` is `true`, then `mbox.recv_timeout()` must be called in `__lumen_builtin_receive_wait` so that the `cursor` is reset and the timer must be canceled.  This means that only `ReceiveState::Received` should happen in `__lumen_builtin_receive_done`.