Fix "no tests to run" for wasm-pack run
Our versions of wasm-pack and wasm-bindgen got out-of-sync, so that the test runner couldn't find the tests anymore, so update all the wasm dependencies to fix it.

As `async` is stable in Rust now, switch from `wasm_bindgen_test(async)` to `wasm_bindgen_test` on an `async` function.  This eliminates the need for the `futures` crate and is more up-to-date with the community best practices.