feat: implement backtraces, refactor exception system
NOTE: This does not fully remove the use of `source` in the exceptions
system, but I believe we can eventually migrate all uses of `source` to
the same trace system implemented here. To do so will require a flag
to indicate whether or not to capture the rust frames, as well as some
additional work to handle rust frames during symbolication, but is
relatively minimal effort. The nice thing is that this will allow traces
to be fully uniform whether generated from the runtime, or thrown by the
generated code.