Fix HeapFragment::new_tuple_from_slice
Fixes #555

The layout for a HeapFragment for a tuple needs to include the space for all elements recursively also because `clone_to_heap` will be called on each element too.  `Tuple::layout_for(elements)` only included the size of the element terms after the header, which only works for immediate terms.