refactor: use stricter approach to types
- No longer treat Box<T> like pointers, but as tagged pointers
- Instead rely on Ptr<T>/Ref<T> for raw pointer-like uses (such as with
MallocOp)
- Treat BooleanType like a real boolean, and provide a dedicated
constant for it
- Instead of defaulting to always casting concrete types to TermType and
then at each use where we need a concrete type cast back; propagate the
concrete type as far as possible and only perform conversion to opaque
term types when required.
- Reimplement CmpEq/CmpNeq in such a way that comparisons can more often
be performed at compile-time, or unboxed, rather than always needing to
call into the runtime. This not only helps optimize performance, but
code size as well.
- Since function calls almost always require opaque terms, we special
case the lowering of operands in CallOp/InvokeOp to cast arguments to
opaque term type if needed. Since lowered intrinsics do not use our
CallOp, this does mean that if we have intrinsics that make calls to the
runtime that we will need to do similar lowerings there.

The primary motivation for this was the discovery that due to always
casting to term type, we had some pathological cases where stupid simple
things like the following:

    case Msg of
      {'DOWN', _, process, _} ->
        ...
    end

Which simplified to something like:

    if is_tuple(Msg) of
      if elem(Msg, 0) == 'DOWN' andalso elem(Msg, 2) == process of
        ...
      end
    end

The result of which generated the following awful IR:

- Perform `eir.is_type` on `Msg`, branch to block with extracted
elements
- Perform `eir.cmp.eq` on first element, branch to next block
- Perform `eir.cmp.eq` on third element, branch to next block
- Cast i1 from both comparisons above to atom type
- Create constant atom for true
- Perform `eir.cmp.eq` on each casted i1 against true
- Perform `eir.logical.and` on the result of _that_ to get a single i1
representing the overall result of the second `if`
- Each of the above in turn might lower to something that casts its
operands

When it should be:

- Perform `eir.is_type` on `Msg`, branch to block with extracted
elements
- Perform `eir.cmp.eq` on first and third elements, branch to next block
- Perform `eir.logical.and` on the result of both comparisons

The second one looks similar, but ultimately lowers to just a couple
instructions (two `icmp` and an `and`). Even better, because we're
operating on primitives, we can often eliminate conditional branches
where we know the result due to constant propagation.

Another problem I identified is that the casts hid a lot of obvious
errors because we essentially treated everything as an opaque value,
this made it very difficult for the compiler to catch validation issues
before they made it to LLVM, usually hitting an assert somewhere in
there. At least now when there are issues, we should catch it with nicer
diagnostics. Unfortunately we may have more frequent compilation
failures in the near term while we work out the bugs that have snuck
into the code so far.

An additional benefit is that this sets us up a lot better for pushing
more type information down into codegen, improving optimizations and
generated code size.