Support variadic logic and and or from EIR
LogicalOperator has a Vec of operands now instead of just rhs and lhs.  The operands vec is converted to a vec of ValueRefs and passed to teh MLIRBuildLogic*Op C functions as a ptr and len like the argc and argc.  When converting to MLIR Operations though, the variadics are removed and instead a chain of LogicalAndOp or LogicalOrOp are constructed, so that the operation can continue to subclass BinaryOperator.