refactor: compiler v2
This commit contains the in-progress work that has occurred since the
last release of the compiler. It is a major rewrite of the compiler
internals, as well as some parts of the runtime, in order to enable the
following:

* Ability to target Wasm without stack switching primitives
* Reimplement the middle tier of the compiler (i.e. replaces EIR with a
  new middle-tier compiler infrastructure). This uses a more traditional
  lowering approach than the one attempted in EIR, and provides some
  benefits we've long wanted (such as the ability to generate a readable
  textual form)
* Provide a richer pass infrastructure that can be used across tiers of
  the compiler.
* Allow most compiler work to be done in Rust. Previously almost all of
the codegen work was based on C++, but this rewrite moves most of this
into Rust, and only requires C++ for things that we don't have an FFI
interface for yet, or which are required to integrate with various
MLIR/LLVM APIs. This should make it much easier to contribute for folks
that are comfortable with Rust, but not C++.

In addition, some significant changes to how codegen works and thus how
it integrates with the runtime:

* The Erlang calling convention now leverages multi-value return
* Exceptions are in the process of being transitioned away from
  C++-style exceptions to using a return-value flag. On native
  architectures we can use CPU flags for this, and on Wasm we have
  native multi-value returns.
* Yielding on Wasm requires continuations, unlike most other
  architectures where we can use stack switching. A specific pass is
  dedicated to performing the transformation necessary to make this work
  as if stack switching was in effect.
* We are shifting to a model where we construct atom and dispatch tables
  at link-time, rather than during compilation of a module. This is
  necessary in order to support linking with other objects as well as
  generation of dynamic libraries. The specific mechanics of this are
  still up in the air.