fix: enable codegen of raw assembly
This commit includes a variety of changes that bring the compiler to the
point where it can generate raw assembly and link executables again,
though the latter requires fresh testing.

Among the changes is a rewrite of how the atom and function dispatch
tables work to rely on the linker to merge all entries in the respective
tables into a single array-per-table which is then used to initialize
them during startup. This ensures we can compile artifacts independently
and link them together correctly, though further testing needs to be
done to understand how dynamically-linked libraries will tie into this
scheme. For now, statically linked executables will work as expected.

A number of other improvements to the IR builder infra are also
included.