wip: reimplement term encoding/representation and implementation
The general aim here is to drastically simplify the implementation of
our runtime code to make iteration faster and to drop dead weight from
many of our experiments in the past. In the immediate term, the plan is
to get the essentials functional and integrated with the compiler before
reintroducing many of the bifs.