fix: address memory safety violation in syntax_core
Function layouts contain intrusive linked lists based on typed arenas
for allocating nodes. Because we had not specified how the drop was
performed, and due to the ordering of the fields, the nodes were being
dropped before the linked list, which could result in use-after-free.
Resolving this is done in two ways, firstly, by addressing the ordering
of the fields so that the list is dropped before the nodes, and secondly
by implementing Drop to ensure that the list is fast-cleared rather than
walked to drop each node, since dropping the arena will drop all nodes
at once.