fix: a variety of codegen/runtime bugs
The fixes in this commit are almost too numerous to list, but are
largely following up on initial implementation of various features with
comprehensive testing and related fixes. A few of the things addressed:

- Binary matching/construction
- Term encoding/decoding/type checking
- Closure creation/application
- Improved diagnostics reporting for parsing issues and other middle
  tier errors/warnings.

There are a handful of improvements to the codegen pipeline as well,
moving more things into Rust, removing some boilerplate C++ code (with
more to be removed in the near future). Further unifying and expanding
on the handling of types from the middle tier through to codegen to make
things such as function signatures at the high level lower correctly to
LLVM IR.