feat: implement syntax_pp crate
This crate builds on the AST defined in firefly_beam to handle parsing
Erlang Abstract Format into the same structure that is parsed from BEAM
Abst/Dbgi chunks. This can then be translated to syntax_erl or
syntax_core.

This commit does not perform lowering to the middle-tier yet, that is
coming in subsequent commits. However, we do wire up the new parser to
the driver and its handling of application resource files.