feat: document and improve compiler interface
Up until now, we have not documented how to work with the compiler once
built, it was largely up to interested individuals to read the help
output and experiment, so it was about time to address this.

Not only does this commit introduce documentation on the various ways you can
interact with the compiler; it also improves the compiler defaults, and
extends the behavior to better support common scenarios such as
compiling multiple applications into a single lib/executable, and
formalizes the way we process input paths.

From an internals perspective, the following is now the way the compiler
understands inputs, outputs, and their relationships:

* If no paths are provided, the compiler assumes the current working
  directory is either a standard Erlang application, or a directory
  containing Erlang sources to be compiled.
* If a single path is provided:
  * If that path is a directory, it treats that directory as the root
    application being compiled
  * If that path is a file, it infers application metadata from that
    file, and treats that application as the root application.
* If multiple paths are provided:
  * The root application is constructed using flags/inference based on
  the current working directory.
  * Each directory is treated as a separate application, and treated as
  an implicit dependency of the root application.
  * All direct file paths are gathered together and treated as members
  of the root application.
* Compiler flags can be used to override/embellish application metadata
  inference. This is currently limited to the application name, version,
  and start module.
* If an application has a start module defined in its metadata, then
  the compiler will default to producing an executable. Otherwise, it
  will default to producing a static library.
* The default artifact type (bin vs lib) can be overridden with compiler
  flags, and in the case of libs, the type of lib can be overridden as
  well (using --dynamic or --static).
* All inputs compiled in a single session of the compiler will be output
  into a single file (executable or library), with the exception of
  debugging artifacts which are still per-module.

To run a completely barebones Erlang executable (i.e. with a custom
init exported as `init:boot/1`):

    firefly compile -C no_default_init --bin init.erl

The resulting binary will be unable to use anything that relies on the
default standard library infrastructure (unless it has been otherwise
provided by the custom init).