fix: improper handling of nested binary match trees
The issue here is that we were returning a mutated match context when a
match was successful, but if a chain of matches failed after one or more
successful matches, the next match chain would receive a match context
that had been modified up to the last successful match in the first
chain. This is obviously incorrect, and the solution is to ensure that
we only return mutated _copies_ of the match context.

This commit applies that fix, but in a less than ideal way - namely, if
a match is successful, it clones the new match context on to the process
heap so that our current calling convention is preserved. The downside
of this is that it incurs an unnecessary allocation on the process heap,
when we should just be able to return the new context on the stack.
However our calling convention currently provides no safe mechanism for
returning structures that are larger than the machine pointer size, and
the only exceptions to this are our multi-value return convention, and
the few variants of it for specific compiler intrinsics. I'd like to
define a safe mechanism by which we can express special types like the
match context and have them be lowered correctly on each platform, but
also ensure that we catch any instances where such a type is improperly
treated like a term. At the moment, we haven't defined any machinery to
support this kind of thing, and rather than define yet another special
case for match contexts, I'm punting on that until I have time to sort
out the calling convention details after some other runtime work in the
queue.