fix: interaction between gc and term/heap fragments
There were a few issues discovered in both the garbage
collector and in the implementation of term and heap fragments.

* The heap iterator when hitting holes was skipping an extra word
on accident due to the iterator `next` implementation advancing
the iterator on each call. If the caller then "skipped" some number
of words or bytes, the extra word was not taken into account and thus
could result in skipping over terms that needed to be visited, such as
ref-counted pointers.
* In a few places, we were incorrectly managing reference-counted terms.
When constructing term fragments and building lists, we were creating
unnecessary clones in some cases and not cleaning them up properly, a
problem for ref-counted terms. When sweeping terms, we were not
incrementing the reference count for ref-counted terms, which could then
be incorrectly freed when the source heap was later reaped.

I'm still not 100% confident in how we're managing ref-counted terms in
general, we'll have to pay close attention to this in the future.