feat: improve implementation of jump tables in bytecode
Previously, we would issue a series of Breq instructions
each of which would decode the argument term, compare against
an immediate for a match, then conditionally branch to the target.
Now, the argument is decoded only once, and the target branch is
resolved in a single instruction. The overall code size is reduced
as well.

Additionally, I've gone through and documented all of the
bytecode ops with their semantics, and assumptions about the
runtime implementation in each.