fix: incorrect catch handler installation/removal
In some situations, namely tail calling a function from within a catch
handler, the catch handler would not be uninstalled, and could be
unwound to again if the callee also raised an error, resulting in an
infinite loop.

The problem was that we were installing catch handlers on entry to a
try or catch expression, but only removing them when exited normally,
and ultimately when returning from a function all of the catch handlers
in that frame would be removed. This was incorrect and resulted in the
problem observed above.

I initially tried to address this in the behavior of the landing_pad
opcode, but due to the semantic differences between try/catch and bare
catch, this lacked sufficient information to do the right thing in both
cases. Instead, I've fixed this in the compiler itself, by generating
the appropriate end_catch ops when entering the catch handler itself.